+++ tests/images/builder/utils.py
@@ -1,0 +1,31 @@
+import logging
+from pathlib import Path
+import subprocess
+import sys
+from typing import List
+
+logging.basicConfig(level=logging.DEBUG)
+log = logging.getLogger(__name__ if __name__ != "__main__" else "convert-image")
+
+BUILD_DIR = "/tmp"
+HOST_PATH = Path("/host")
+
+
+def build_path(path: Path) -> Path:
+    """
+    Convert a host filesystem path to its corresponding path inside the Docker container.
+
+    The Docker container mounts the host's root filesystem at /host, so this function
+    transforms absolute host paths like '/home/user/file.txt' to container paths
+    like '/host/home/user/file.txt'.
+
+    Args:
+        path: Absolute or relative path on the host filesystem
+
+    Returns:
+        Path that can be used inside the Docker container to access the same file
+
+    Example:
+        build_path(Path("/home/user/config.yaml")) -> Path("/host/home/user/config.yaml")
+    """
+    return HOST_PATH / path.absolute().relative_to(Path("/"))
